UPSTREAM: ARM: keystone2: move update of the phys-to-virt constants into generic code
Make the init_meminfo function return the offset to be applied to the
phys-to-virt translation constants.  This allows us to move the update
into generic code, along with the requirements for this update.

This avoids platforms having to know the details of the phys-to-virt
translation support.

Acked-by: Santosh Shilimkar <ssantosh@kernel.org>
Tested-by: Murali Karicheri <m-karicheri2@ti.com>
Signed-off-by: Russell King <rmk+kernel@arm.linux.org.uk>

[Amit: Fixes keystone_defconfig build. Ideally it should have been picked
       up before Change-Id: Iecdf1ce57dcb898996e8b9acedefd3c2b98735f8
       ("BACKPORT: ARM: re-implement physical address space switching")]
(cherry picked from commit c8ca2b4b29282605698968d15667939b23e208e2)
Signed-off-by: Amit Pundir <amit.pundir@linaro.org>